lambda/requirements.txt
PyNaCl
boto3
